requirements.txt
pandas
requests
openai
nltk
torch
transformers
scikit-learn
bertopic
numpy
pydantic
matplotlib
python-dotenv